#654: moved install plugins from run to postInstall adjusted vscode and intellij tests added --force to vscode plugin installation
